SERVER-4985: use 64-bit integer for connection and thread id.

+++ src/mongo/util/goodies.h
@@ -30,7 +30,7 @@ namespace mongo {
      * @return if this name has an increasing counter associated, return the value
      *         otherwise 0
      */
-    unsigned setThreadName(const char * name);
+    long long setThreadName(const char * name);
     string getThreadName();
 
     template<class T>
